// packages/davidia-components/src/plots/AnyPlot.d.ts
interface MP_NDArray {
    /** If it is an n-dimensional array */
    nd: boolean;
    /** The data type */
    dtype: string;
    /** The shape of the data */
    shape: number[];
    /** The data */
    data: ArrayBuffer;
}
// ...
interface BatonProps {
    /** The universally unique identifier (uuid) of the client */
    uuid: string;
    /** The uuid of the current baton holder */
    batonUuid: string | null;
    /** The other uuids */
    others: string[];
    /** If client holds baton */
    hasBaton: boolean;
    /** Handles baton request */
    requestBaton: () => void;
    /** Approves passing baton to client with given uuid */
    approveBaton: (s: string) => void;
}
/**
 * The props for the `PlotSelectionProps` component.
 * @interface {object} PlotSelectionProps
 * @member {(selection: SelectionBase | null, broadcast?: boolean, clear?: boolean) => void} addSelection - Handles adding selection.
 * @member {SelectionBase[]} selections - The selections.
 * @member {BatonProps} [batonProps] - The baton props.
 */
interface PlotSelectionProps {
    /** Handles adding selection */
    addSelection?: (selection: SelectionBase | null, broadcast?: boolean, clear?: boolean) => void;
    /** The selections */
    selections: SelectionBase[];
    /** The baton props */
    batonProps?: BatonProps;
}
// ...
interface SurfacePlotProps extends PlotSelectionProps {
    /** The data values for the surface plot */
    values: NdArray<TypedArray>;
    /** The domain of the surface data */
    domain: Domain;
    /** The axes parameters */
    axesParameters: DAxesParameters;
    /** The colour scale type */
    surfaceScale: ColorScaleType;
    /** The colour map (optional) */
    colourMap?: ColorMap;
}
// ...
interface DAxesParameters {
    /** The x-axis label (optional) */
    xLabel?: string;
    /** The y-axis label (optional) */
    yLabel?: string;
    /** The x-axis scale type (optional) */
    xScale?: AxisScaleType;
    /** The y-axis scale type (optional) */
    yScale?: AxisScaleType;
    /** The x-axis values (optional) */
    xValues?: NdArray<TypedArray>;
    /** The y-axis values (optional) */
    yValues?: NdArray<TypedArray>;
    /** The plot title (optional) */
    title?: string;
}